+++ lib/src/animation/content/fill_content.dart
@@ -1,7 +1,5 @@
 import 'dart:ui';
-
 import 'package:vector_math/vector_math_64.dart';
-
 import '../../l.dart';
 import '../../lottie_drawable.dart';
 import '../../lottie_property.dart';