+++ lib/src/animation/content/shape_modifier_content.dart
@@ -3,4 +3,6 @@ import 'content.dart';
 
 abstract class ShapeModifierContent extends Content {
   ShapeData modifyShape(ShapeData shapeData);
+
+  void addUpdateListener(void Function() listener);
 }